Add python project files, README, and Makefile
This establishes a baseline set of files for the project. It contains
the following:

    1. Makefile
    2. README.md
    3. docs/ directory
    4. requirements.txt
    5. setup.py
    6. wordsearch/ module

The Makefile is included to make running certain tasks easier. The
README.md now contains a rough draft of how to set up the project,
install and uninstall.

The `docs/` directory contains files generated by Sphinx with a few
minor changes to the configuration to include Python docstring
automatically.

The setup.py and requirements.txt are necessary for installing the
module itself, as well as fetching the necessary dependencies.

As for the module itself, I only included enough code to test the
installation procedure, and to make sure I can import the module and run
tests on it; I expect for the contents of these files to change
significantly as the project progresses.

Closes: #6