Add function to parse a puzzle file
I added a function that takes a open file descriptor and returns the
parsed word list and puzzle as a tuple containing a list and matrix
(two-dimensional array).

This implementation is clearly naive (in that it doesn't actually read
anything from the file); it will need to be adjusted to handle different
puzzle files.