Refactor parse_puzzle so that it handles the input puzzle
I modified the parse_puzzle function so that it actually reads the data
from the input puzzle, parses each line, and produces the list of words
along with a matrix containing the puzzle board.

The approach used here assumes that the first line of the file is the
word list, and that commas are used to separate each word (as well as
separate the characters in the puzzle board).