Add test for ensuring default values from parse_puzzle
This adds a small sanity check just to make it more explicit that I
intend for parse_puzzle to return a tuple of the form (None, []) if the
file received is empty. The reasoning here is that if you give nothing
to parse, then there can't possible be any result.

I could have made this raise an error, but it would complicate the
implementation, and returning empty values is just as effective. This,
of course, puts the burden on the user of this function to check the
result.