Add Puzzle class
According to the requirements, the input file is supposed to contain a
puzzle board that is square, and the minimum size that a word can be is
two characters. So logically, the board cannot have a height and width
smaller than two.

The constructor enforces this invariant so that the class methods can
operate on these assumptions safely.