Refactor size() by adding width and height properties
Since size is really just the height and width as a tuple, I refactored
it so that it just returns a tuple of two new properties that return the
height and width respectively. This makes it so that internally I don't
have to split the tuple to get the width and height individually.