Modify get_valid_moves to consider distance
It only makes sense to consider searching in a particular direction if
moving that many spaces in a given direction moves across valid
positions. By multiplying the direction tuple by some distance, it
projects the position multiple spaces in that direction.

For example, if you are at position (0,0) and you want check if moving
three spaces to the right is still within the bounds of the matrix, then
take the tuple representing each direction and multiply it by a
distance. The tuple (0,1) is to the right, but multiplying both terms by
a distance measure of 3 results in (0,3). Adding both terms to the
original position yields (0,3), which is still within bounds.