Add get_direction method to Puzzle class
It is useful to be able to determine the direction that a given point
(origin) would need to move in to reach another point (target) when you
know both the origin and target. By knowing the direction, expressed as
a tuple such as (0,1) for 'right' and (1,0) for 'down', the moves
necessary to get to the point 'target' from the point 'origin' is the
direction tuple added to the origin n number of times, where n is the
distance from the origin to the target.

For example, if you have an origin point (0,0) and a target point (0,3),
then the distance between these points is three. The direction, 'right',
expressed as a tuple is (0,1). By adding (0,1) to (0,0) three times, we
arrive at (0,3), i.e. the target point.