Add get_characters method to Puzzle class
Building on the get_direction and get_valid_moves methods, the
get_characters method takes two points (an origin and target returned
from get_valid_moves), determines what direction the target is in with
respect to the origin point, and collects the characters along the path
from the origin to the target (inclusive). The collected characters are
returned along with their positions so that it becomes possible to do
string comparison against words from our word list.