Implement main method
The main method is now able to take advantage of the tools we've built
up to easily read in the data, solve the puzzle, and output the
formatted output.

I included an integration test to ensure that the Puzzle class could
consume data from a file and solve the puzzle, and an end-to-end test to
make sure that running the program from the command-line produces the
expected output in the terminal.

Closes: #1, #2, #3, #4, #5, #21